Suggested edit for #A glimmer of hope
The original one: 
It was a boring night, and nothing seemed out of the ordinary. The only sound that could be heard was the machinery that propelled the space elevator. From time to time, a transport chamber would arrive at or depart from the station, generating a distinct sound. When a chamber arrived, it emitted a sound similar to that of a steam engine, and when it hit the platform, a deep, loud noise followed.

Although these events disrupted the tranquility of the night, they paled in comparison to the turmoil generated by the space elevator's operations in the past. Those were better times for space commerce; nowadays, these operations only run at a minimum level. In the past, the station required around a hundred people to operate, whereas now, only a handful of people are needed for a typical day's operation, and they work in different parts of the station.

Some workers at the station spent their hours in complete isolation, with no interaction with others. Despite strict restrictions on the planet, including a complete smoking ban, one worker smoked and watched the figures produced by his cigarette, assigning names and familiar shapes to the small forms rising from it until they disappeared. He complained about the limitations of the body and wished he were like gas or smoke, able to easily escape from the confinement that others called the workplace.

After a few minutes, the sensation of the cigarette in his mouth was just a memory. But he was grateful that the chemicals he had just absorbed were now in his brain, helping him to get through the remaining hours of work. The task was straightforward: check the transport chambers and ensure that all the specified merchandise was there. As someone with access to these chambers before anyone else, he could receive and put smuggling. Obtaining cigarettes this way was risky, as the punishment for smuggling was death, but life was a burden to him. From his perspective, cigarettes were worth the risk.


After hours of tedious work, the burden became unbearable, and suddenly an idea flashed through his mind. What if he got into one of the chambers and escaped? He knew that all the products were destined for a neighboring planet, and once there, he could start a new one without the absurd limitations that his native planet's leaders had imposed. Life was beginning to make sense again.